Fixing the Video Canvas
Last problem seems to be caused by client and host assigning the values to the server. Of course this will fail, because only the host can write to the server's networkvariable, but the client does this in the code currently as well. This needs to be looked a bit more in depth but after that I doubt that the video player needs more fixing.

The slider bop or whatever it's called now also tracks correctly between the host and the clients. Of course this will need to be tested on a two different pcs that is the volume correctly adjusted to the client and server.

This is purposely done in a wrong way. What I mean by that is that you shouldn't really even try to sync the volume between all the users but it's there to demonstrate exactly why that is the case. The code definitely will need some comments but I'll be taking care of that later.