Table Problem continues
There's definitely a bug in unity's end regarding this.

For some reason when the client is supposed to pick up the table, the isKinematic bool is set to true permanently and the user can't move the object at all.

It's a bug with ownership and the isKinematic bool which I have to look more into what is causing it and how to fix it. I'll give one more shot before I'll close the pc for today.