Closer to the answer
Now the user is able to move the table on the server, however I need to handle the event of setting the kinematic object better. After the player now releases the table it will toggle the kinematic states in wrong way however if other object is selected and the player now tries to move the first table, the table moves once again.