The table is now fixed
By adjusting the values in network physics interactable script you might  make it work better, as for now you can't spam the table or the kinematic value gets messed up. However if managed with some patience (waiting 1 second before touching the table again) the user is able to finally move the table even if it's spawned by the host.

The fact that this isn't a problem with the pucks or the mallet is still mystery for me. I'm not sure exactly why the table seems to be the only problem.

The problem was fixed in the last commit, this commit is just a mass change to table so i can update that last commit worked perfectly.